+++ include/H2DE/H2DE_engine.h
@@ -16,6 +16,7 @@
 #include <filesystem>
 #include <iostream>
 #include <string>
+#include <map>
 #include <unordered_map>
 #include <vector>
 
@@ -39,7 +40,6 @@ private:
 
     std::unordered_map<std::string, SDL_Texture*> textures;
     std::unordered_map<std::string, Mix_Chunk*> sounds;
-    std::unordered_map<std::string, TTF_Font*> fonts;
 
     std::vector<H2DE_GraphicObject*> graphicObjects;
     std::optional<H2DE_SClick*> click = std::nullopt;
@@ -90,12 +90,6 @@ private:
      */
     void importSound(const fs::path& sound);
     /**
-     * Imports a font from a ttf file
-     * \param font ttf file
-     * \since H2DE-1.0.0
-     */
-    void importFont(const fs::path& font);
-    /**
      * Indicates that an asset has been loaded
      * \since H2DE-1.0.0
      */
@@ -139,12 +133,6 @@ private:
      * \since H2DE-1.0.0
      */
     void renderCircle(H2DE_GraphicObject* g);
-    /**
-     * Renders a text on the renderer
-     * \param g the text to render
-     * \since H2DE-1.0.0
-     */
-    void renderText(H2DE_GraphicObject* g);
 
     /**
      * Simulates a click on a graphic element