+++ src/coin.cpp
@@ -1,0 +1,11 @@
+#include "item.h"
+
+// INIT
+Coin::Coin(Game* g, LevelPos p) : Item(g, p, "coin.png") {
+
+}
+
+// CLEANUP
+Coin::~Coin() {
+    
+}